Backend/requirements.txt
django==4.1.3
djangorestframework==3.14.0
djangorestframework-simgplejwt==5.1.0
django-cors-headers==3.13.0
PyJWT==2.6.0
Pillow==9.3.0
mysqlclient==2.1.1